requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
svgwrite==1.4.3
scikit-image==0.21.0
scikit-learn==1.3.0
